Refactor : NotificationToast 적용 및 Header -> HeaderLayout으로 변경
