+++ src/types.ts
@@ -27,3 +27,9 @@ export interface ProductWithUI extends Product {
   description?: string;
   isRecommended?: boolean;
 }
+
+export interface Notification {
+  id: string;
+  message: string;
+  type: "error" | "success" | "warning";
+}
